[src] Move composers
